Moved DatePickerDialog to ui/utils/Common
